check for instanceof ConstantDataSetDescriptor in loadDataSet, and don't start a DataRequestThread in this case.  This fixes the flashiness problem in the slicers.
perhaps a better way to do this is for the dataSetDescriptor to indicate
that it can produce its dataset in animation-interactive time, so that
the overhead of creating the thread can be avoided generally.

There is a potential problem in this implementation in that the
"Rebinning Dataset" message is not displayed in the progress panel
for ConstantDataSetDescriptors.

svn path=/dasCore/trunk/; revision=413