introduction of the InputStreamMeter, which monitors InputStreams on a system level.
MeteredInputStream is function is now subsumed by the InputStreamMeter, which
meters a set of input streams to better measure and limit the connection speed. (The MeteredInputStream could only meter the one stream, so there was no way to
get the aggregate speed, which better represents the potential of the resource.

bugfix: Also, this gets rid of the lifecycle bug where one MeteredInputStream
would be told to meter two input streams, causing an annoying RuntimeException.

svn path=/dasCore/trunk/; revision=2186